requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
exceptiongroup==1.2.2
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.2.0
incremental==24.7.2
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
priority==2.0.0
py-mini-racer==0.6.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
taskgroup==0.0.0a4
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
zope.interface==7.0.1
